财务管理

+++ jk_entity/src/main/java/top.greathead.jk.entity/Invoice.java
@@ -18,7 +18,7 @@ public class Invoice extends BaseEntity {
 	private String scNo;			//packing.getExportNos S/C就是报运的合同号
 	private String blNo;			
 	private String tradeTerms;			
-	private Double state;			//0草稿 1已上报
+	private Long state;			//0草稿 1已上报
 	private String createBy;			
 	private String createDept;			
 	private Date createTime;
@@ -59,11 +59,11 @@ public class Invoice extends BaseEntity {
 		this.tradeTerms = tradeTerms;
 	}
 
-	public Double getState() {
+	public Long getState() {
 		return state;
 	}
 
-	public void setState(Double state) {
+	public void setState(Long state) {
 		this.state = state;
 	}
